begin converting string method plugin to westpa 2.0
