+++ src/westpa/westext/wess/wess_driver.py
@@ -4,9 +4,9 @@ import operator
 import numpy as np
 
 import westpa
-from westpa.yamlcfg import check_bool
-from westpa.kinetics import RateAverager
-from westext.wess.ProbAdjust import prob_adjust
+from westpa.core.yamlcfg import check_bool
+from westpa.core.kinetics import RateAverager
+from westpa.westext.wess.ProbAdjust import prob_adjust
 
 EPS = np.finfo(np.float64).eps
